requirements.txt
absl-py==1.3.0
accelerate @ git+https://github.com/huggingface/accelerate.git@180ef7c41576bc55cb1265fcdeb7da287e95c9b7
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
arviz==0.15.1
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
bcbio-gff==0.6.6
beautifulsoup4==4.11.2
bio==1.3.3
biopython==1.79
biothings-client==0.2.6
bleach==6.0.0
blessed==1.20.0
brotlipy==0.7.0
cachetools==5.2.0
certifi @ file:///croot/certifi_1683875369620/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.3
comm==0.1.2
conda==22.11.1
conda-package-handling @ file:///croot/conda-package-handling_1672865015732/work
conda_package_streaming @ file:///croot/conda-package-streaming_1670508151586/work
cons==0.4.5
contourpy==1.0.5
cryptography @ file:///croot/cryptography_1673298753778/work
cycler==0.11.0
datasets==2.13.1
debugpy==1.6.6
decorator==5.1.1
deepspeed==0.9.4
defusedxml==0.7.1
dill==0.3.6
etuples==0.3.8
evaluate==0.4.0
executing==1.2.0
fastjsonschema==2.16.2
fastprogress==1.0.3
filelock==3.9.0
flatbuffers==22.9.24
fonttools==4.37.4
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.4.0
gast==0.4.0
genomic_embeddings==0.0.4
gensim==3.8.3
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpustat==1.1
grpcio==1.49.1
h5netcdf==1.1.0
h5py==3.7.0
hjson==3.1.0
huggingface-hub==0.16.4
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==5.0.0
install==1.3.5
ipdb==0.13.13
ipykernel==6.24.0
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kaleido==0.2.1
keras==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lab==7.2
libclang==14.0.6
libmambapy @ file:///home/conda/feedstock_root/build_artifacts/mamba-split_1667627019743/work/libmambapy
lit==16.0.1
llvmlite==0.39.1
logical-unification==0.4.6
loralib==0.1.1
mamba @ file:///home/conda/feedstock_root/build_artifacts/mamba-split_1667627019743/work/mamba
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
miniKanren==1.0.3
mistune==2.0.5
mpmath==1.3.0
multidict==6.0.4
multipledispatch==0.6.0
multiprocess==0.70.14
mygene==3.2.2
mysql-connector-python==8.0.33
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numcodecs==0.7.3
numpy==1.22.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==11.525.112
nvidia-ml-py3==7.352.0
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
peft @ git+https://github.com/huggingface/peft.git@06fd06a4d2e8ed8c3a253c67d9c3cb23e0f497ad
pexpect==4.8.0
pickle5==0.0.11
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.0.0
plotly==5.14.1
pluggy @ file:///tmp/build/80754af9/pluggy_1648024445381/work
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.9
Pygments==2.14.0
pymc==5.3.1
pynndescent==0.5.8
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
pytensor==2.11.3
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.4
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rouge-score==0.1.2
rsa==4.9
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
safetensors==0.3.1
scikit-learn==0.24.1
scipy==1.10.1
seaborn==0.10.0
Send2Trash==1.8.0
simplejson==3.18.3
six==1.16.0
smart-open==6.2.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
sympy==1.11.1
tenacity==8.2.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
tensorrt==8.5.3.1
termcolor==2.0.1
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.2
tomli==2.0.1
toolz @ file:///croot/toolz_1667464077321/work
torch==2.0.1
torchvision==0.15.1
tornado==6.2
tqdm @ file:///croot/tqdm_1679561862951/work
traitlets==5.9.0
transformers @ git+https://github.com/huggingface/transformers.git@35eac0df75c692c5b93c12f7eaf3279cab8bd7ce
triton==2.0.0
txt2tags==3.7
typing_extensions==4.4.0
tzdata==2023.3
umap==0.1.1
umap-learn==0.5.3
uri-template==1.2.0
urllib3 @ file:///croot/urllib3_1686163155763/work
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wrapt==1.14.1
xarray==2023.4.2
xarray-einstats==0.5.1
xgboost==1.3.3
xxhash==3.2.0
yarl==1.9.1
zipp==3.9.0
zstandard @ file:///opt/conda/conda-bld/zstandard_1663827383994/work
